:art: - Moved /executions/delete endpoint to api file
