+++ packages/cli/src/Server.ts
@@ -1897,81 +1897,6 @@ class App {
 
 		this.app.use(`/${this.restEndpoint}/executions`, executionsController);
 
-		// Delete Executions
-		// INFORMATION: We use POST instead of DELETE to not run into any issues
-		// with the query data getting to long
-		this.app.post(
-			`/${this.restEndpoint}/executions/delete`,
-			ResponseHelper.send(async (req: ExecutionRequest.Delete): Promise<void> => {
-				const { deleteBefore, ids, filters: requestFilters } = req.body;
-
-				if (!deleteBefore && !ids) {
-					throw new Error('Either "deleteBefore" or "ids" must be present in the request body');
-				}
-
-				const sharedWorkflowIds = await getSharedWorkflowIds(req.user);
-				const binaryDataManager = BinaryDataManager.getInstance();
-
-				// delete executions by date, if user may access the underyling worfklows
-
-				if (deleteBefore) {
-					const filters: IDataObject = {
-						startedAt: LessThanOrEqual(deleteBefore),
-					};
-
-					if (filters) {
-						Object.assign(filters, requestFilters);
-					}
-
-					const executions = await Db.collections.Execution.find({
-						where: {
-							workflowId: In(sharedWorkflowIds),
-							...filters,
-						},
-					});
-
-					if (!executions.length) return;
-
-					const idsToDelete = executions.map(({ id }) => id.toString());
-
-					await Promise.all(
-						idsToDelete.map(async (id) => binaryDataManager.deleteBinaryDataByExecutionId(id)),
-					);
-
-					await Db.collections.Execution.delete({ id: In(idsToDelete) });
-
-					return;
-				}
-
-				// delete executions by IDs, if user may access the underyling worfklows
-
-				if (ids) {
-					const executions = await Db.collections.Execution.find({
-						where: {
-							id: In(ids),
-							workflowId: In(sharedWorkflowIds),
-						},
-					});
-
-					if (!executions.length) {
-						LoggerProxy.error('Failed to delete an execution due to insufficient permissions', {
-							userId: req.user.id,
-							executionIds: ids,
-						});
-						return;
-					}
-
-					const idsToDelete = executions.map(({ id }) => id.toString());
-
-					await Promise.all(
-						idsToDelete.map(async (id) => binaryDataManager.deleteBinaryDataByExecutionId(id)),
-					);
-
-					await Db.collections.Execution.delete(idsToDelete);
-				}
-			}),
-		);
-
 		// ----------------------------------------
 		// Executing Workflows
 		// ----------------------------------------
